+++ src/Amadevus.RecordGenerator.Generators/Names.cs
@@ -20,5 +20,10 @@
 
         public const string ToolName = "Amadevus.RecordGenerator";
         public const string GeneratedCodeAttribute = "System.CodeDom.Compiler.GeneratedCodeAttribute";
+
+        public const string SystemCollectionsGenericNamespace = "System.Collections.Generic";
+        public const string EqualityComparerName = "EqualityComparer";
+        public const string SystemNamespace = "System";
+        public const string IEquatableName = "IEquatable";
     }
 }