+++ src/Amadevus.RecordGenerator.Generators/RecordDescriptor.cs
@@ -7,13 +7,16 @@ namespace Amadevus.RecordGenerator.Generators
 {
     internal partial class RecordDescriptor
     {
-        public RecordDescriptor(Features features, TypeSyntax Type, SyntaxToken TypeIdentifier, ImmutableArray<Entry> Entries, TypeDeclarationSyntax TypeDeclaration)
+        public RecordDescriptor(Features features, TypeSyntax Type, SyntaxToken TypeIdentifier, ImmutableArray<Entry> Entries, TypeDeclarationSyntax TypeDeclaration, Location TypeDeclarationLocation, ISymbol Symbol, SemanticModel SemanticModel)
         {
             this.Type = Type;
             this.TypeIdentifier = TypeIdentifier;
             this.Entries = Entries;
             this.TypeDeclaration = TypeDeclaration;
             Features = features;
+            this.TypeDeclarationLocation = TypeDeclarationLocation;
+            this.Symbol = Symbol;
+            this.SemanticModel = SemanticModel;
         }
 
         public Features Features { get; }
@@ -26,9 +29,15 @@ namespace Amadevus.RecordGenerator.Generators
 
         public TypeDeclarationSyntax TypeDeclaration { get; }
 
+        public Location TypeDeclarationLocation { get; }
+
+        public ISymbol Symbol { get; }
+
+        public SemanticModel SemanticModel { get; }
+
         internal abstract class Entry
         {
-            public Entry(SyntaxToken Identifier, TypeSyntax Type, PropertyDeclarationSyntax PropertySyntax)
+            public Entry(SyntaxToken Identifier, TypeSyntax Type, PropertyDeclarationSyntax PropertySyntax, ISymbol TypeSymbol)
             {
                 this.Identifier = Identifier;
                 this.Type = Type;
@@ -36,6 +45,7 @@ namespace Amadevus.RecordGenerator.Generators
                 var id = (string)Identifier.Value;
                 var camelized = char.ToLowerInvariant(id[0]) + id.Substring(1);
                 IdentifierInCamelCase = SyntaxFactory.Identifier(CSharpKeyword.Is(camelized) ? "@" + camelized : camelized);
+                this.TypeSymbol = TypeSymbol;
             }
 
             public SyntaxToken Identifier { get; }
@@ -45,11 +55,13 @@ namespace Amadevus.RecordGenerator.Generators
             public TypeSyntax Type { get; }
 
             public PropertyDeclarationSyntax PropertySyntax { get; }
+
+            public ISymbol TypeSymbol { get; }
         }
 
         internal class SimpleEntry : Entry
         {
-            public SimpleEntry(SyntaxToken Identifier, TypeSyntax Type, PropertyDeclarationSyntax PropertySyntax) : base(Identifier, Type, PropertySyntax)
+            public SimpleEntry(SyntaxToken Identifier, TypeSyntax Type, PropertyDeclarationSyntax PropertySyntax, ISymbol TypeSymbol) : base(Identifier, Type, PropertySyntax, TypeSymbol)
             {
             }
         }