+++ src/Amadevus.RecordGenerator.Generators/RecordGenerator.cs
@@ -1,12 +1,12 @@
-﻿using CodeGeneration.Roslyn;
-using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.CSharp;
-using Microsoft.CodeAnalysis.CSharp.Syntax;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using CodeGeneration.Roslyn;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 
 namespace Amadevus.RecordGenerator.Generators
 {
@@ -26,11 +26,12 @@ namespace Amadevus.RecordGenerator.Generators
 
             if (context.ProcessingNode is ClassDeclarationSyntax classDeclaration)
             {
-                var descriptor = classDeclaration.ToRecordDescriptor(features);
+                var descriptor = classDeclaration.ToRecordDescriptor(features, context.SemanticModel);
                 generatedMembers = generatedMembers
                     .AddRange(
                         GenerateRecordPartials(descriptor)
                         .Where(x => x != null));
+                foreach (var diagnostic in GenerateDiagnostics(descriptor)) progress.Report(diagnostic);
             }
             return Task.FromResult(generatedMembers);
 
@@ -43,7 +44,17 @@ namespace Amadevus.RecordGenerator.Generators
                 yield return RecordPartialGenerator.Generate(descriptor, cancellationToken);
                 yield return BuilderPartialGenerator.Generate(descriptor, cancellationToken);
                 yield return DeconstructPartialGenerator.Generate(descriptor, cancellationToken);
-                yield break;
+                if (descriptor.Symbol.IsSealed)
+                {
+                    yield return ObjectEqualsGenerator.Generate(descriptor, cancellationToken);
+                    yield return EquatableEqualsPartialGenerator.Generate(descriptor, cancellationToken);
+                }
+                yield return OperatorEqualityPartialGenerator.Generate(descriptor, cancellationToken);
+            }
+
+            IEnumerable<Diagnostic> GenerateDiagnostics(RecordDescriptor descriptor)
+            {
+                foreach (var diagnostic in EqualityUsageAnalyzer.GenerateDiagnostics(descriptor)) yield return diagnostic;
             }
         }